python/requirements.txt
Flask
requests
beautifulsoup4
mysql-connector